+++ Aula 4/desafio006.py
@@ -1,2 +1,2 @@
-n = int(input("Digite um número: "))
+n = float(input("Digite um número: "))
 print(f"O dobro do número digitado: {n * 2};\nO triplo é: {n * 3};\nE a raiz quadrada é: {n**(1/2)}")
