+++ Aula 4/desafio008.py
@@ -1,4 +1,4 @@
 print("")
-valor_em_metros = int(input("Digite o valor em metros: "))
+valor_em_metros = float(input("Digite o valor em metros: "))
 print("")
 print(f"O valor em centímetros é: {valor_em_metros * 100} cm;\nO valor em milímetros: {valor_em_metros * 1000} mm")
