+++ Aula 4/desafio011.py
@@ -1,5 +1,5 @@
 print("")
-largura = int(input("Digite a largura da parede: ")) 
-altura = int(input("Digite a altura da parede: "))
+largura = float(input("Digite a largura da parede: ")) 
+altura = float(input("Digite a altura da parede: "))
 print("")
 print(f"A área da parede é: {largura *  altura} m²;\nSerá necessário usar: {(largura * altura) / 2} litros de tinta")
